# Make Qp.integer_ring() faster.